lang: remove &mut tokens, remove tks lifetime & Enumerate source in Scanner
It makes more sense for the Scanner to own the Vec of Tokens while it's actually parsing and then move ownership when
the Scanner is ending it's lifetime on source()'s stack. This allows us to remove the old lifetime, 'tks, associated with the mutable reference.
Wrapping the source Chars in a Enumerate<Chars> is good for if we ever need to know the index of where we are in the Chars iterator,
such as creating a Token::Identifier or Token::Number.